requirements.txt
requests
requests
openai>=1.10.0
azure-core>=1.29.5
python-dotenv>=1.0.0